some initial refactoring of sidebar components
